Remove Client Mixin
